Closing GH-2

+++ src/main/java/org/weatherScrape/entitiy/Day.java
@@ -13,7 +13,7 @@ public class Day extends WeatherData {
 
     private int uvIndex;
 
-    public Day(int temperature, int cloudCover, int windGust, int windSpeed, String windDirection, int thunderProbability, int precipitation, int precipitationProbability, String statement, int uvIndex) {
+    public Day(int temperature, int cloudCover, int windGust, int windSpeed, String windDirection, int thunderProbability, double precipitation, int precipitationProbability, String statement, int uvIndex) {
         super(temperature, cloudCover, windGust, windSpeed, windDirection, thunderProbability, precipitation, precipitationProbability, statement);
         this.uvIndex = uvIndex;
     }
